Compiling version, it lacks the skipper and proper tests
